Pulse doesn't work as much as we thought... High delta makes it get a pretty darn good outcome, but with a little delta we get terrible results. But it runs rl fast (pruning WAY too much)

+++ vns.py
@@ -1431,7 +1431,7 @@ def appendToSchedSteps(stringInfo, schedOrRoute, routeIndex = None):
             
 def main():
     # print "********** Main **********"
-    stoppingCondition = int(sys.argv[2])
+    stoppingCondition = float(sys.argv[2])
     #run vns on file specified in user input
     f = sys.argv[1]
     f = os.path.realpath(f)
